+++ spec/support/active_record_init.rb
@@ -4,6 +4,8 @@ require "active_record"
 require "activerecord-jdbc-adapter" if defined? JRUBY_VERSION
 require "activerecord-jdbcsqlite3-adapter" if defined? JRUBY_VERSION
 
+ActiveRecord::Base.raise_in_transactional_callbacks = true unless
+  ActiveRecord::VERSION::MAJOR >= 5
 ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
 
 ActiveRecord::Schema.define do